Add queries from database history notebook
The database history notebook contains 12 queries; one for each table.
This study contains one query, which is the union of the 12 queries. One
query is executed by one SQL Runner action and outputs one CSV file,
which reduces the number of actions and CSV files we need to work with
in downstream actions.

The unioned queries in this study are, on the whole, equivalent to the
non-unioned queries in the database history notebook. There are two
exceptions. The first contains `CodedEvent` and `Appointment`, where
`DATETIME`s are converted to `DATE`s. This is for consistency with other
non-unioned queries. The second contains `SGSS_Positive`/`SGSS_Negative`
and `SGSS_AllTests_Positive`/`SGSS_AllTests_Negative`, where counts for
each table are output, rather than counts for the union of both SGSS_*
and `SGSS_AllTests_*` tables. This is for convenience, as it means each
unioned query follows the same pattern.