Run tests with `opensafely exec`
This means that we run tests in a Python container, so the test
environment and the production environment are the same.

We remove pytest and pytest-cov from requirements.dev.in because we
won't be running tests locally. However, pytest is a dependency of
sqlfluff, so it hasn't been removed from requirements.dev.txt. This
dependency is most annoying.

We add opensafely to requirements.dev.* so that `opensafely exec` is on
the path.