Add script to generate more realistic dummy rows
SQL Runner can't generate dummy rows, so we committed some to get things
moving; the file contained column headers, but no more. More realistic
dummy rows are necessary for running the pipeline locally and in CI, so
this script generates them.

As before, we commit a file: this is necessary for running the pipeline
in CI. An alternative would be to add an action to *project.yaml*, but
an action called `generate_dummy_rows` would look odd on OS Jobs. We
move the file from *analysis/dummy_rows.csv.gz* to
*output/generate_dummy_rows/dummy_rows.csv.gz* to clearly separate
analysis from output.