Aggregate event counts and apply SDC functions
The Statistical Disclosure Control (SDC) functions are located in
`analysis.sdc` to signal their importance and to make them easier to
reason about.

Rather than call `redact_le` and `round_to_nearest` directly, we create
partial functions to make it clear to a reader of `analysis.aggregate`
what we are redacting to and what we are rounding to, without them
having to read `analysis.sdc`.