Render the report as an HTML file
Here, we add the `render_report` action to render the report as an HTML
file using the Jinja templating engine. We use the same method as
opensafely-core/interactive-templates; that is, we encode the plots of
aggregated event counts using Base64, so that we don't have to include
them in the list of outputs in project.yaml.