Make query consistent with report text
Several tables referenced by the query contain dates in the far future
(e.g. the year 9999), probably because these dates represent null
values. The report text, which was copied from the notebook text, says
that the figures represent event activity from 2016-01-01 to the date
the report was run. We can avoid having to handle dates in the far
future in downstream actions by making the query consistent with the
report text, and returning event activity to the date the report was
run.