Add plots from 2020
This commit adds plots from 2020, which feature in the original database
builds notebook (plots from 2016 feature in the original database
history notebook). As the 2020 png files have the same names as the 2016
png files, they are name-spaced by passing an `--output` argument. (And
because we are passing this argument, we can delete `config.FROM_DATE`.)

We maintain the structure of the existing TOC because we expect this
structure to be most useful to users; that is, first you look for the
table you're interested in, then you look for the date-range you're
interested in. We place the narrower date-ranges above the wider
date-ranges, again, because we expect this structure to be most useful
to users. (If not, we can simply reorder the positional arguments to
`render_report.group_plots`.)