Use random walk to generate event counts
Previously, event counts were drawn from a discrete uniform distribution
(`rng.integers`). As such, the event count for a given day was unrelated
to the event count for the previous day; this produced noisy plots. A
random walk is a better model, as the event count for a given day is
related to the event count for the previous day; this produces plots
with clear trends.

As well as being visually more appealing, plots with clear trends should
help us spot errors.