Fix `aggregate`
When the resampling function produced a series of floats, as was the
case when it was the mean, values greater than 7.0 but less than 7.5
would be rounded down to 5. Now, values are rounded to the nearest
integer and the series is cast to an `int`, so values in this range will
be either redacted (e.g. 7.1 -> 7 -> 0) or rounded up to 10 (e.g. 7.5 ->
8 -> 10).

Closes #93